refactor: move legacy (gw) class definition types to the common crate
