+++ crates/common/src/lib.rs
@@ -14,6 +14,7 @@ use pathfinder_crypto::Felt;
 use primitive_types::H160;
 use serde::{Deserialize, Serialize};
 
+pub mod class_definition;
 pub mod consts;
 pub mod event;
 pub mod hash;
