+++ crates/gateway-types/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod class_definition;
 pub mod class_hash;
 pub mod error;
 pub mod reply;
