+++ crates/gateway-types/src/request.rs
@@ -119,100 +119,14 @@ impl From<BlockNumberOrTag> for pathfinder_common::BlockId {
     }
 }
 
-pub mod contract {
-    use std::fmt;
-
-    use fake::{Dummy, Fake, Faker};
-    use pathfinder_common::{ByteCodeOffset, EntryPoint};
-    use pathfinder_crypto::Felt;
-    use serde_with::serde_as;
-
-    #[derive(Copy, Clone, Debug, serde::Deserialize, serde::Serialize, PartialEq, Hash, Eq)]
-    #[serde(deny_unknown_fields)]
-    pub enum EntryPointType {
-        #[serde(rename = "EXTERNAL")]
-        External,
-        #[serde(rename = "L1_HANDLER")]
-        L1Handler,
-        #[serde(rename = "CONSTRUCTOR")]
-        Constructor,
-    }
-
-    impl fmt::Display for EntryPointType {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            use EntryPointType::*;
-            f.pad(match self {
-                External => "EXTERNAL",
-                L1Handler => "L1_HANDLER",
-                Constructor => "CONSTRUCTOR",
-            })
-        }
-    }
-
-    #[serde_as]
-    #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, PartialEq)]
-    #[serde(deny_unknown_fields)]
-    pub struct SelectorAndOffset {
-        pub selector: EntryPoint,
-        #[serde_as(as = "OffsetSerde")]
-        pub offset: ByteCodeOffset,
-    }
-
-    #[derive(serde::Deserialize, serde::Serialize)]
-    #[serde(untagged)]
-    pub enum OffsetSerde {
-        HexStr(Felt),
-        Decimal(u64),
-    }
-
-    impl serde_with::SerializeAs<ByteCodeOffset> for OffsetSerde {
-        fn serialize_as<S>(source: &ByteCodeOffset, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: serde::Serializer,
-        {
-            use serde::Serialize;
-
-            Felt::serialize(&source.0, serializer)
-        }
-    }
-
-    impl<'de> serde_with::DeserializeAs<'de, ByteCodeOffset> for OffsetSerde {
-        fn deserialize_as<D>(deserializer: D) -> Result<ByteCodeOffset, D::Error>
-        where
-            D: serde::Deserializer<'de>,
-        {
-            use serde::Deserialize;
-
-            let offset = OffsetSerde::deserialize(deserializer)?;
-            let offset = match offset {
-                OffsetSerde::HexStr(felt) => felt,
-                OffsetSerde::Decimal(decimal) => Felt::from_u64(decimal),
-            };
-            Ok(ByteCodeOffset(offset))
-        }
-    }
-
-    impl<T> Dummy<T> for SelectorAndOffset {
-        fn dummy_with_rng<R: rand::prelude::Rng + ?Sized>(_: &T, rng: &mut R) -> Self {
-            Self {
-                selector: Faker.fake_with_rng(rng),
-                offset: ByteCodeOffset(Felt::from_u64(rng.gen())),
-            }
-        }
-    }
-
-    /// Descriptor of an entry point in a Sierra class.
-    #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, Dummy)]
-    #[serde(deny_unknown_fields)]
-    pub struct SelectorAndFunctionIndex {
-        pub selector: EntryPoint,
-        pub function_idx: u64,
-    }
-}
-
 pub mod add_transaction {
     use std::collections::HashMap;
 
+    use pathfinder_common::class_definition::{
+        EntryPointType,
+        SelectorAndFunctionIndex,
+        SelectorAndOffset,
+    };
     use pathfinder_common::{
         AccountDeploymentDataElem,
         CasmHash,
@@ -227,7 +141,6 @@ pub mod add_transaction {
     use pathfinder_serde::{CallParamAsDecimalStr, TransactionSignatureElemAsDecimalStr};
     use serde_with::serde_as;
 
-    use super::contract::{EntryPointType, SelectorAndFunctionIndex, SelectorAndOffset};
     use super::{CallParam, ContractAddress, Fee, TransactionSignatureElem};
     use crate::reply::transaction::{DataAvailabilityMode, ResourceBounds};
 
@@ -465,11 +378,10 @@ pub mod add_transaction {
         }
 
         mod byte_code_offset {
+            use pathfinder_common::class_definition::SelectorAndOffset;
             use pathfinder_common::macro_prelude::*;
             use pathfinder_common::ByteCodeOffset;
             use pathfinder_crypto::Felt;
-
-            use crate::request::contract::SelectorAndOffset;
 
             #[test]
             fn with_hex_offset() {
